ct: move level zero read path into level zero package
This includes the bits from cloud_topics/read_path as well as the level
zero reader implementation that was co-located with the frontend.
