+++ src/v/cloud_topics/data_plane_impl.cc
@@ -16,10 +16,10 @@
 #include "cloud_topics/ephemeral_cluster_services.h"
 #include "cloud_topics/level_zero/batcher/batcher.h"
 #include "cloud_topics/level_zero/read_pipeline.h"
+#include "cloud_topics/level_zero/reader/fetch_request_handler.h"
 #include "cloud_topics/level_zero/reconciler/reconciler.h"
 #include "cloud_topics/level_zero/throttler/throttler.h"
 #include "cloud_topics/level_zero/write_pipeline.h"
-#include "cloud_topics/read_path/fetch_request_handler.h"
 #include "model/fundamental.h"
 #include "model/record_batch_reader.h"
 #include "storage/types.h"
