+++ src/v/cloud_topics/level_zero/reader/README.md
