+++ src/v/cloud_topics/level_zero/reader/fetch_request_handler.cc
@@ -8,14 +8,14 @@
  * https://github.com/redpanda-data/redpanda/blob/master/licenses/rcl.md
  */
 
-#include "cloud_topics/read_path/fetch_request_handler.h"
+#include "cloud_topics/level_zero/reader/fetch_request_handler.h"
 
 #include "base/unreachable.h"
 #include "cloud_topics/dl_placeholder.h"
 #include "cloud_topics/level_zero/event_filter.h"
 #include "cloud_topics/level_zero/read_request.h"
+#include "cloud_topics/level_zero/reader/materialized_extent_reader.h"
 #include "cloud_topics/logger.h"
-#include "cloud_topics/read_path/materialized_extent_reader.h"
 #include "model/record.h"
 #include "model/record_batch_reader.h"
 #include "model/timeout_clock.h"
