+++ src/v/cloud_topics/level_zero/reader/fetch_request_handler.h
