+++ src/v/cloud_topics/level_zero/reader/materialized_extent_reader.h
