+++ src/v/cloud_topics/level_zero/reader/reader.h
