+++ src/v/cloud_topics/level_zero/reader/tests/BUILD
@@ -9,7 +9,7 @@ redpanda_test_cc_library(
         "materialized_extent_fixture.h",
         "mocks.h",
     ],
-    include_prefix = "cloud_topics/read_path/tests",
+    include_prefix = "cloud_topics/level_zero/reader/tests",
     visibility = ["//visibility:public"],
     deps = [
         "//src/v/base",
@@ -24,7 +24,7 @@ redpanda_test_cc_library(
         "//src/v/cloud_topics:object_utils",
         "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/read_path:materialized_extent",
+        "//src/v/cloud_topics/level_zero/reader:materialized_extent",
         "//src/v/container:fragmented_vector",
         "//src/v/model",
         "//src/v/model/tests:random",
@@ -49,8 +49,8 @@ redpanda_cc_gtest(
         "//src/v/cloud_topics:extent_meta",
         "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/read_path:materialized_extent",
-        "//src/v/cloud_topics/read_path/tests:materialized_extent_fixture",
+        "//src/v/cloud_topics/level_zero/reader:materialized_extent",
+        "//src/v/cloud_topics/level_zero/reader/tests:materialized_extent_fixture",
         "//src/v/container:fragmented_vector",
         "//src/v/model",
         "//src/v/ssx:sformat",
@@ -70,8 +70,8 @@ redpanda_cc_gtest(
     deps = [
         "//src/v/base",
         "//src/v/cloud_topics:types",
-        "//src/v/cloud_topics/read_path:materialized_extent_reader",
-        "//src/v/cloud_topics/read_path/tests:materialized_extent_fixture",
+        "//src/v/cloud_topics/level_zero/reader:materialized_extent_reader",
+        "//src/v/cloud_topics/level_zero/reader/tests:materialized_extent_fixture",
         "//src/v/container:fragmented_vector",
         "//src/v/model",
         "//src/v/ssx:sformat",
@@ -92,8 +92,8 @@ redpanda_cc_gtest(
         "//src/v/base",
         "//src/v/cloud_topics:types",
         "//src/v/cloud_topics/level_zero:read_pipeline",
-        "//src/v/cloud_topics/read_path:fetch_handler",
-        "//src/v/cloud_topics/read_path/tests:materialized_extent_fixture",
+        "//src/v/cloud_topics/level_zero/reader:fetch_handler",
+        "//src/v/cloud_topics/level_zero/reader/tests:materialized_extent_fixture",
         "//src/v/container:fragmented_vector",
         "//src/v/model",
         "//src/v/ssx:sformat",
