+++ src/v/cloud_topics/level_zero/reader/tests/materialized_extent_fixture.cc
