+++ src/v/cloud_topics/level_zero/reader/tests/mocks.h
