+++ src/v/kafka/data/BUILD
@@ -39,7 +39,7 @@ redpanda_cc_library(
         "//src/v/cloud_topics:placeholder",
         "//src/v/cloud_topics/frontend",
         "//src/v/cloud_topics/frontend:errc",
-        "//src/v/cloud_topics/frontend:level_zero_reader_impl",
+        "//src/v/cloud_topics/level_zero/reader",
         "//src/v/kafka/protocol",
         "//src/v/kafka/server:errors",
         "//src/v/model",
