+++ src/v/kafka/data/cloud_topic_partition.cc
@@ -17,7 +17,7 @@
 #include "cloud_topics/extent_meta.h"
 #include "cloud_topics/frontend/errc.h"
 #include "cloud_topics/frontend/frontend.h"
-#include "cloud_topics/frontend/level_zero_reader_impl.h"
+#include "cloud_topics/level_zero/reader/reader.h"
 #include "cluster/partition.h"
 #include "cluster/rm_stm.h"
 #include "cluster/types.h"
